Add a function and docs for fetching Alps GPS
The function is in the `ensaio.v1` module. The module contains a
`pooch.Pooch` with the Zenodo DOI as a source and destination in a
`ensaio/v1` cache folder. That way, we can add a `v2` later when there
are breaking changes in the data but still keep the `v1` data available,
regardless the version of Ensaio itself. Add a gallery example with this
data to test that it works.