Version datasets individually  (#18)
Instead of using the entire data bundle and versioning all datasets based on
the module name, we can now version them individually. To do so, we split up
the data archives into separate repositories and Zenodo releases of
https://github.com/fatiando-data. This means that functions won't have to be
repeated and updating one dataset doesn't mean copying all of the others along
with it (since the collection would be new). Versions are now specified as a
required `version` argument in all `fetch_*` functions. When updating a
dataset, functions for the others don't need to be repeated in a new module.
Also only need 2 environment variables for setting the cache location and the
data source (instead of 2 per version). A downside is that we can no longer
accept a variable to set custom data source URL since each dataset has a
different one. The new environment variable only sets fetching from GitHub or
not. There was considerable refactoring of the code to make this work.