A logo for Ensaio
The design is of course inspired by David Bowie. The story behind this:
The name Ensaio was suggested by my wife after a long brainstorming
session while we were reading the book "Little people big dreams: My
first David Bowie" to my daughter. The name came to her as we saw Bowie
rehearsing (*ensaio* in Portuguese) with loads of different styles and
colors. The colors are from the Harmonica logo which is a reference to
the red-while-blue colormap often used for gravity and magnetic data.