Convert tutorials from sphinx-gallery to plain rst
The tutorials have much more text than Python code so rendering them in
sphinx-gallery means writing a bunch of rst content as comments in the
`.py` files, which is very awkward. Use jupyter-sphinx instead so we can
have the tutorials as plain rst files with code blocks executed by the
extension. The output looks pretty much the same with some slight
differences in style (had to tweak the CSS to match). Another advantage
is that the "Edit on GitHub" links from sphinx-book-theme now work for
these (they don't for the gallery output).