Update the Caribbean bathymetry data to v2 (#54)
The new version has fewer points (smaller area), removed the very dense
surveys, removed the surveys with systematic errors. All of this to help
reduce the file size and number of points to something more manageable.
Version 1 took too long to just plot. Add a test function that downloads
the v2 datasets as well. For now, make this manual since this is the
only dataset with v2.